fix: rename kebab-case files to snake_case and resolve eslint errors
Rename 6 files from kebab-case to snake_case to satisfy filename-naming-convention rule.
Fix DEFAULT_ID → defaultId in board.test.ts and replace `as any` with `as unknown as T` in session_events.test.ts.
