+++ src/routes/issues/+page.svelte
@@ -5,7 +5,7 @@
 	import { useVersionControl } from '$lib/modules/version-control';
 	import { useActions } from '$lib/modules/actions';
 	import { useNotifications, NOTIFICATION_DOT_COLORS } from '$lib/modules/notifications';
-	import { findNotificationDotColor } from '$lib/modules/notifications/notification-helpers.js';
+	import { findNotificationDotColor } from '$lib/modules/notifications/notification_helpers.js';
 	import { useSessions } from '$lib/modules/sessions';
 	import type { Issue, CreateIssueRequest, UpdateIssueRequest } from '$lib/modules/issues';
 	import type { PrunableIssue } from '$lib/types/generated';
